Deprecate the StatusChecksProperties interface and add AbstractStatusChecksProperties class insead.
